backend/requirements.txt
flask
flask-cors
boto3
python-dotenv
pdfplumber
pinecone-client
sentence-transformers
requests